# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed: one entry in errors[] per invalid field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a task (API-TSK-001).",
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page, 1–100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter to these statuses; absent = no filter; `status=` or an unknown value → 400.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TaskStatus"
              }
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort order; `title` compares case-insensitively. Ties break on `id` descending.",
            "schema": {
              "default": "updatedAt:desc",
              "type": "string",
              "enum": [
                "updatedAt:desc",
                "updatedAt:asc",
                "dueDate:asc",
                "dueDate:desc",
                "title:asc",
                "title:desc"
              ]
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Filter to a single project; no value or foreign/trashed project → empty items.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive \"contains\" filter on the title; empty = no filter.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed: one entry in errors[] per invalid field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The caller's tasks, optionally filtered by project and status (API-TSK-002).",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "operationId": "TasksController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One task (API-TSK-003).",
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed: one entry in errors[] per invalid field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change the given fields (API-TSK-004).",
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Moved to the trash."
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Move the task to the trash (API-TSK-005).",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}/restore": {
      "post": {
        "operationId": "TasksController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Restored, or already active."
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Take the task out of the trash (API-TSK-006).",
        "tags": [
          "Tasks"
        ]
      }
    }
  },
  "info": {
    "title": "PM4 API",
    "description": "The PM4 backend API (see ../specs/03-api).",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Dot path of the field, e.g. `address.city`."
          },
          "message": {
            "type": "string",
            "description": "What is wrong with it."
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "The HTTP status code.",
            "example": 400
          },
          "type": {
            "type": "string",
            "description": "`<WEB_APP_URL>/errors/<slug>`, e.g. `https://pm4.angelovski.top/errors/validation`."
          },
          "title": {
            "type": "string",
            "description": "Short summary of the status, e.g. `Validation failed`."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation."
          },
          "errors": {
            "description": "Only on a `400` validation error: one entry per failed constraint.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "status",
          "type",
          "title",
          "detail"
        ]
      },
...
      "LeadUserDto": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "description": "The user's current display name (D2)."
          }
        },
        "required": [
          "avatarUrl",
          "id",
          "displayName"
        ]
      },
      "ProjectLeadDto": {
        "type": "object",
        "properties": {
          "kind": {
            "enum": [
              "user",
              "text"
            ],
            "type": "string"
          },
          "user": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LeadUserDto"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "`user.displayName` for a user lead, else the saved text name."
          }
        },
        "required": [
          "kind",
          "user",
          "name"
        ]
      },
...
      "TaskStatus": {
        "type": "string",
        "enum": [
          "upcoming",
          "in-progress",
          "completed"
        ],
        "description": "Status of the task; absent defaults to `upcoming`. `null` → 400."
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true,
            "description": "≤ 2000 chars; empty after trimming → `null`.",
            "maxLength": 2000
          },
          "externalLink": {
            "type": "string",
            "nullable": true,
            "description": "Absolute `http(s)://` URL, ≤ 500 chars; empty after trimming → `null`.",
            "maxLength": 500,
            "format": "uri"
          },
          "projectLeadUserId": {
            "type": "string",
            "nullable": true,
            "description": "A registered user as the lead; excludes `projectLeadName`.",
            "format": "uuid"
          },
          "projectLeadName": {
            "type": "string",
            "nullable": true,
            "description": "A free-text lead name, ≤ 100 chars; empty after trimming → `null`.",
            "maxLength": 100
          },
          "status": {
            "description": "Status of the task; absent defaults to `upcoming`. `null` → 400.",
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              }
            ]
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "description": "Work date in `YYYY-MM-DD`; empty after trimming → `null`.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "projectId": {
            "type": "string",
            "description": "The project that contains this task.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "1–200 chars after trimming.",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "projectId",
          "title"
        ]
      },
      "RefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "deleted"
        ]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "externalLink": {
            "type": "string",
            "nullable": true
          },
          "projectLead": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectLeadDto"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              }
            ]
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "description": "Work date, `YYYY-MM-DD`"
          },
          "id": {
            "type": "string"
          },
          "project": {
            "$ref": "#/components/schemas/RefDto"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 UTC"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601 UTC"
          }
        },
        "required": [
          "description",
          "externalLink",
          "projectLead",
          "status",
          "dueDate",
          "id",
          "project",
          "title",
          "createdAt",
          "updatedAt"
        ]
      },
      "TaskListResponseDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponseDto"
            }
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "items"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "The project to move the task to; not `null`.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "1–200 chars after trimming; not `null`.",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "≤ 2000 chars; empty after trimming → `null`.",
            "maxLength": 2000
          },
          "externalLink": {
            "type": "string",
            "nullable": true,
            "description": "Absolute `http(s)://` URL, ≤ 500 chars; empty after trimming → `null`.",
            "maxLength": 500,
            "format": "uri"
          },
          "projectLeadUserId": {
            "type": "string",
            "nullable": true,
            "description": "A registered user as the lead; excludes `projectLeadName`.",
            "format": "uuid"
          },
          "projectLeadName": {
            "type": "string",
            "nullable": true,
            "description": "A free-text lead name, ≤ 100 chars; empty after trimming → `null`.",
            "maxLength": 100
          },
          "status": {
            "description": "Status of the task; not `null`.",
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              }
            ]
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "description": "Work date in `YYYY-MM-DD`; empty after trimming → `null`.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          }
        }
      }
    }
  }
}